playbooks/ops/migrate_images_to_swift/container-src/requirements.txt
adal==1.2.2
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
dictdiffer==0.8.0
enum34==1.1.6
futures==3.3.0
google-auth==1.6.3
idna==2.8
ipaddress==1.0.22
iso8601==0.1.12
Jinja2==2.10.1
keystoneauth1==3.14.0
kubernetes==8.0.1
MarkupSafe==1.1.1
oauthlib==3.0.2
openshift==0.8.0
os-service-types==1.7.0
pbr==5.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
python-string-utils==0.6.0
python-swiftclient==3.7.0
PyYAML==3.13
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.ordereddict==0.4.13
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
six==1.12.0
stevedore==1.30.1
urllib3==1.25.3
websocket-client==0.56.0
